# Use CTE for optimized address transactions query